fix: Updated Str parameter to utilize Strinct functionality. Removed Strinct parameter.
