+++ src/param.js
@@ -2,7 +2,6 @@ const Str = require('./params/str');
 const RegEx = require('./params/regex');
 const Email = require('./params/email');
 const UUID = require('./params/uuid');
-const Strinct = require('./params/strinct');
 const IsoDate = require('./params/iso-date');
 const Date = require('./params/date');
 const Bool = require('./params/bool');
@@ -21,7 +20,6 @@ module.exports.Str = (...args) => new Str(...args);
 module.exports.RegEx = (...args) => new RegEx(...args);
 module.exports.Email = (...args) => new Email(...args);
 module.exports.UUID = (...args) => new UUID(...args);
-module.exports.Strinct = (...args) => new Strinct(...args);
 module.exports.IsoDate = (...args) => new IsoDate(...args);
 module.exports.Date = (...args) => new Date(...args);
 module.exports.Bool = (...args) => new Bool(...args);
