+++ test/handler.js
@@ -82,7 +82,6 @@ module.exports.param = api.wrap('POST param', [
     maxItems: 3,
     minItems: 1
   }),
-  api.Strinct('strinctParam', 'query', { required: false }),
   api.FieldsParam('fieldsParam', 'query', { required: false, fields: 'id' }),
   api.Number('numberParam', 'query', { required: false }),
   api.NumberList('numberListParam', 'query', { required: false }),
